PR review - moved common code to BaseIntegrationTest class
